requirements.txt
aioredis==2.0.1
amqp==5.1.1
appdirs==1.4.4
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
bs4==0.0.1
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
channels==3.0.5
chardet==5.0.0
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
constantly==15.1.0
cryptography==37.0.4
cssselect==1.1.0
daphne==3.0.2
Deprecated==1.2.13
distlib==0.3.5
Django==4.0.6
django-celery-beat==2.3.0
django-celery-results==2.4.0
django-ckeditor==6.5.1
django-crispy-forms==1.14.0
django-js-asset==2.0.0
django-timezone-field==5.0
fake-useragent==0.1.11
feedparser==6.0.10
filelock==3.7.1
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.12.0
incremental==21.3.0
kombu==5.2.4
lxml==4.9.1
msgpack==1.0.4
numpy==1.23.1
packaging==21.3
pandas==1.4.3
parse==1.19.0
Pillow==9.2.0
platformdirs==2.5.2
prompt-toolkit==3.0.30
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyee==8.2.2
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyppeteer==1.0.2
pyquery==1.4.3
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.1
redis==4.3.4
requests==2.28.1
requests-html==0.10.0
service-identity==21.1.0
sgmllib3k==1.0.0
simplejson==3.17.6
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tqdm==4.64.0
Twisted==22.4.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.3.0
tzdata==2022.1
urllib3==1.26.11
vine==5.0.0
virtualenv==20.16.2
w3lib==1.22.0
wcwidth==0.2.5
websockets==10.3
whitenoise==6.2.0
wrapt==1.14.1
yahoo-fin==0.8.9.1
zipp==3.8.1
zope.interface==5.4.0
